requirements.txt
spotipy==2.23.0
yt-dlp==2023.12.30
mutagen==1.47.0
customtkinter==5.2.2
pillow==10.2.0
requests==2.31.0 